Refactored recognizer.ts into multiple files
